+++ lib/smart_form_fields.dart
@@ -1,5 +1,6 @@
 library;
 
+export 'src/animation/smart_error_animation.dart' show SmartErrorAnimation;
 export 'src/form/smart_form.dart' show SmartForm, SmartFormState;
 export 'src/form/smart_form_controller.dart' show SmartFormController;
 export 'src/form/smart_form_key.dart' show SmartFormKey;
@@ -10,3 +11,4 @@ export 'src/fields/smart_form_field.dart'
 export 'src/fields/smart_text_field.dart' show SmartTextField;
 export 'src/validation/smart_async_validator.dart' show SmartAsyncValidator;
 export 'src/validation/smart_validator.dart' show SmartValidator;
+export 'src/validation/smart_validators.dart' show SmartValidators;